% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Example dataset: cancer}
\format{
A data frame with 475 rows and 13 variables:
\describe{
  \item{x}{An integer from 1 to 5}
  \item{y}{A lowercase letter}
}
}
\source{
Generated example
}
\usage{
cancer
}
\description{
A dataset with 475 rows and 13 columns.
}
\keyword{datasets}
