% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_derivs.R
\name{constraints_derivs}
\alias{constraints_derivs}
\title{Derivatives of Parameter Constraints}
\usage{
constraints_derivs(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list containing the first constraint derivatives and the block-diagonal
matrix of second constraint derivatives.
}
\description{
\code{constraints_derivs()} is a generic function for obtaining first and
second derivatives of parameter constraints introduced by manifolds and
model transformations.
}
\details{
Compute first and second derivatives associated with constraints on model
parameters and transformed parameters.
}
\seealso{
\code{\link{jacobian}}, \code{\link{hessian}},
\code{\link{vcov}}
}
