% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.step.R
\name{E.step}
\alias{E.step}
\title{Expectation step}
\usage{
E.step(alpha, beta, gamma, data, min.detect, event)
}
\arguments{
\item{alpha}{vector of intercepts where each entry is the intercept for one species/event combination}

\item{beta}{slope vector with one entry per FIB species, which indicates the marginal increase in that species' log-mean for a unit increase in the contamination index}

\item{gamma}{contamination vector with one entry per row of data, indicating the contamination index for the corresponding row}

\item{data}{matrix of FIB counts with one row per sample and one column per FIB species}

\item{min.detect}{minimum detectable level of the FIB assay - counts below this threshold are censored}

\item{event}{vector with one entry per row of data where each entry indicates the event with which the data row is associated}
}
\value{
A matrix of FIB counts, where the censored counts are replaced by their expectations under the current parameters.
}
\description{
Compute the expectations of censored FIB counts, resitricted to be less than the censoring threshold.
}

