% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empathy.R
\docType{data}
\name{empathy}
\alias{empathy}
\title{Example dataset: empathy}
\format{
A data frame with 467 rows and 13 variables:
\describe{
  \item{x}{An integer from 1 to 5}
  \item{y}{A lowercase letter}
}
}
\source{
Generated example
}
\usage{
empathy
}
\description{
A dataset from the tidySEM R package with 467 rows and 13 columns.
The first 6 columns are continuous item responses.
}
\keyword{datasets}
