% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{An  example illustrating the function of packace latent}
\usage{
example()
}
\value{
latent a list as returned by function \code{\link{latent}}
}
\description{
Estimates a latent contamination variable for the storm sewer data, using EM to impute censored values
}

