% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_llca.R
\name{fitted.llcalist}
\alias{fitted.llcalist}
\title{Fitted Probabilities for Lists of Latent Class Models}
\usage{
\method{fitted}{llcalist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"llcalist"} containing fitted
\code{"llca"} and/or \code{"llca_sam"} objects.}

\item{...}{Additional arguments passed to the corresponding
\code{fitted()} methods.}
}
\value{
A named list containing fitted latent class membership probabilities for
each model in \code{object}. The result has class
\code{"fitted.llcalist"}.
}
\description{
Extract fitted prior latent class membership probabilities from an
\code{"llcalist"} object.
}
\seealso{
\code{\link{fitted.llca}}, \code{\link{predict.llca}}
}
