% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{flatten.listlist}
\alias{flatten.listlist}
\title{Flatten a list of lists}
\usage{
flatten.listlist(lst)
}
\arguments{
\item{lst}{A list consisting of lists. These lists should in turn consist of
either a single number, or a named array.}
}
\value{
An array consisting of all the elements of the above lists.
}
\description{
Converts a list of lists to a single list. Useful for parameter
handling.
}
