% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss82.R
\docType{data}
\name{gss82}
\alias{gss82}
\title{Example dataset: gss82}
\format{
A data frame with 1202 rows and 4 variables:
\describe{
  \item{x}{An integer from 1 to 5}
  \item{y}{A lowercase letter}
}
}
\source{
Generated example
}
\usage{
gss82
}
\description{
A dataset from the poLCA R package with 1202 rows and 4 columns. Items are
dichotomous and polytomous.
}
\keyword{datasets}
