% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{hessian}
\alias{hessian}
\title{Hessian Matrix}
\usage{
hessian(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric matrix containing the Hessian of the objective function with
respect to the freely estimated parameters.
}
\description{
\code{hessian()} is a generic function for obtaining the matrix of
second-order derivatives of the objective function with respect to the
freely estimated model parameters.
}
\details{
Compute the Hessian matrix of a fitted latent variable model.
}
\seealso{
\code{\link{vcov}}, \code{\link{jacobian}},
\code{\link{constraints_derivs}}
}
