% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolzingerSwineford1939.R
\docType{data}
\name{HolzingerSwineford1939}
\alias{HolzingerSwineford1939}
\title{Example dataset: HolzingerSwineford1939}
\format{
A data frame with 301 rows and 15 variables:
\describe{
  \item{x}{An integer from 1 to 5}
  \item{y}{A lowercase letter}
}
}
\source{
Generated example
}
\usage{
HolzingerSwineford1939
}
\description{
A dataset from the lavaan R package.
}
\keyword{datasets}
