% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_lcfa.R
\name{information.lcfa}
\alias{information.lcfa}
\title{Information Variance-Covariance Matrix for Confirmatory Factor Models}
\usage{
\method{information}{lcfa}(fit)
}
\arguments{
\item{fit}{A fitted object of class \code{"lcfa"}.}
}
\value{
A list containing the CFA Hessian and variance-covariance matrix.
}
\description{
Information Variance-Covariance Matrix for Confirmatory Factor Models
}
