% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian}
\alias{jacobian}
\title{Jacobian Matrix}
\usage{
jacobian(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric matrix containing derivatives of transformed model parameters.
}
\description{
\code{jacobian()} is a generic function for obtaining the derivatives
associated with parameter transformations in fitted latent variable models.
}
\details{
Compute Jacobian matrices for transformed model parameters.
}
\seealso{
\code{\link{hessian}}, \code{\link{vcov}},
\code{\link{constraints_derivs}}
}
