% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latent.R
\name{latent}
\alias{latent}
\title{Estimate parameters of a latent variable model}
\usage{
latent(data, min.detect, event, specific = NULL, verbose = TRUE)
}
\arguments{
\item{data}{matrix of observed counts, number of rows equal to the number of observations and columns equal to the number of categories.}

\item{min.detect}{minimum limit of detection for the FIB assay - values below this threshold are censored}

\item{event}{vector of event assignment labels, one entry per row of data. Indicates the event to which each row of data belongs.}

\item{specific}{vector of TRUE/FALSE values, one entry per column of data. Each entry indicates whether the corresponding FIB species is human-specific.}

\item{verbose}{should the function provide verbose output about its progress? Defaults to TRUE.}
}
\value{
A list of results
}
\description{
Estimate the parameters of a latent variable model by the method of maximum likelihood, using conjugate gradient descent to maximize likelihood and an EM algorithm to handle censored data.
}

