% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_class.r
\name{latent_fit}
\alias{latent_fit}
\title{Fit a model}
\usage{
latent_fit(obj, ...)
}
\arguments{
\item{obj}{A model object.}

\item{data}{A data frame.}

\item{weights}{An array of non-negative weights of the same length as the
number of rows in \code{data}.}
}
\value{
A list with fitted parameters.
}
\description{
Fits a model to data using weighted maximum likelihood.
}
