% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_flexsurv_ph.r
\name{latent_flexsurv_ph}
\alias{latent_flexsurv_ph}
\title{Flexible spline proportional hazards model}
\usage{
latent_flexsurv_ph(formula, knots, coef_names = NULL)
}
\arguments{
\item{formula}{An analysis formula. The dependent variable is assumed to be a
\link[survival]{Surv} object in data.}

\item{knots}{An array containing all knots, both boundary knots and inner
knots.}

\item{coef_names}{An optional character array of names for the regression
coefficients, excluding the spline coefficients. The latter are fixed as
\code{gamma0}, \code{gamma1} etc.}
}
\value{
A \code{latent_flexsurv_ph} object.
}
\description{
Creates a model object for flexible spline proportional hazards
models.
}
\details{
The following generic functions can be invoked on the model object:
\itemize{
 \item{\link{latent_fit}}{   Fit the model.}
 \item{\link{latent_likelihood}}{   Compute likelihood.}
 \item{\link{latent_loglik_gradient_list}}{   Compute gradients of the
 log-likelihood.}
 \item{\link{latent_loglik_hessian_list}}{   Compute Hessians of the
 log-likelihood.}
 \item{\link{latent_simulate}}{   Simulate data from the model.}
}
}
