% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_class.r
\name{latent_likelihood}
\alias{latent_likelihood}
\title{Compute likelihood}
\usage{
latent_likelihood(obj, ...)
}
\arguments{
\item{obj}{A model object.}

\item{theta}{A list of model parameters.}

\item{data}{A data frame.}
}
\value{
An array with individual likelihoods.
}
\description{
Computes the likelihood for a model object on given data.
}
