% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_class.r
\name{latent_loglik_gradient_list}
\alias{latent_loglik_gradient_list}
\title{Compute likelihood gradients}
\usage{
latent_loglik_gradient_list(obj, ...)
}
\arguments{
\item{obj}{A model object.}

\item{theta}{A list of model parameters.}

\item{data}{A data frame.}
}
\value{
A list with  gradients for the given individuals.
}
\description{
Computes the gradient (vector with first-order derivatives)
of the log-likelihood for a model object on given data.
}
