% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_class.r
\name{latent_loglik_hessian_list}
\alias{latent_loglik_hessian_list}
\title{Compute likelihood Hessians}
\usage{
latent_loglik_hessian_list(obj, ...)
}
\arguments{
\item{obj}{A model object.}

\item{theta}{A list of model parameters.}

\item{data}{A data frame.}
}
\value{
A list with  Hessians for the given individuals.
}
\description{
Computes the Hessian (matrix with second-order derivatives)
of the log-likelihood for a model object on given data.
}
