% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_main.r
\name{latent_main}
\alias{latent_main}
\title{Effect estimation in latent subgroups}
\usage{
latent_main(
  data,
  modelS,
  model0,
  model1,
  do_em = T,
  max_em_iter = 250,
  loglik_conv_tol = 1e-05
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{modelS}{A model object  (e.g. \link{latent_glm}) for subgroup
# membership (dichotomous outcome data).}

\item{model0}{Outcome model for individuals not belonging to the latent
subgroup (untreatable patients).}

\item{model1}{Outcome model for individuals belonging to the latent
subgroup (treatable patients).}

\item{do_em}{If \code{TRUE}, the EM algorithm will be used. Otherwise,
the likelihood is maximized using the general-purpose optimization function
\link{optim} with \code{method = BFGS}.}

\item{max_em_iter}{Maximum number of iterations in the EM algorithm. If the
algorithm doesn't converge within this number of iterations, an error is
raised.}

\item{loglik_conv_tol}{Convergence tolerance criterion for the EM algorithm.
If the log-likelohood increases less than this amount between two successive
iterations, the algorithm is considered to have converged.}
}
\value{
A list with two fields:
\itemize{
 \item{\code{theta}}{   An array \eqn{\theta = (\theta_S, \theta_0, \theta_1)}
 with fitted parameters from all three models.}
 \item{\code{info}}{   The Fisher information matrix, i.e. the negavive
 Hessian of the log-likelihood with respect to the full set of model
 parameters \eqn{\theta} (all three models).}
}
}
\description{
Main function for performing effect estimation in latent
subgroups.
}
\details{
Dichotomous variables should be 0/1 coded. For the latent subgroup,
0 will correspond to untreatable patients, and 1 to treatable ones.

To avoid name conflicts, the variable names in \code{data} should
not start with a dot (\code{.}). The reason is that this function will add
some variables for internal use, all starting with a dot.

For further details on the usage of this function, see the package vignette.
}
