% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_ordinal.r
\name{latent_ordinal}
\alias{latent_ordinal}
\title{Ordinal proportional odds models}
\usage{
latent_ordinal(formula, K, coef_names = NULL)
}
\arguments{
\item{formula}{An analysis formula.}

\item{K}{The number of levels (at least 3) in the ordinal response variable.}

\item{coef_names}{An optional character array of names for the regression
coefficients, excluding the intercepts. The latter are fixed as
\code{alpha1}, \code{alpha2} etc.}
}
\value{
A \code{latent_ordinal} object.
}
\description{
Creates a model object for ordinal models, see the package
vignette for details. The outcome variable should be a factor with 3+ levels.
It need not be \link{ordered} in R, the order of the levels defined the
ordinality.
}
\details{
The following generic functions can be invoked on the model object:
\itemize{
 \item{\link{latent_fit}}{   Fit the model.}
 \item{\link{latent_likelihood}}{   Compute likelihood.}
 \item{\link{latent_loglik_gradient_list}}{   Compute gradients of the
 log-likelihood.}
 \item{\link{latent_loglik_hessian_list}}{   Compute Hessians of the
 log-likelihood.}
 \item{\link{latent_simulate}}{   Simulate data from the model.}
}
}
