% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_flexsurv_ph.r
\name{latent_simulate.latent_flexsurv_ph}
\alias{latent_simulate.latent_flexsurv_ph}
\title{Simulate outcome}
\usage{
\method{latent_simulate}{latent_flexsurv_ph}(obj, theta, data)
}
\description{
Simulates outcomes from a survival model,
see \link{latent_simulate} for details.
}
