% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_glm.r
\name{latent_simulate.latent_glm}
\alias{latent_simulate.latent_glm}
\title{Simulate outcome}
\usage{
\method{latent_simulate}{latent_glm}(obj, theta, data)
}
\description{
Simulates outcomes from a generalized linear model,
see \link{latent_simulate} for details.
}
