% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_linear.r
\name{latent_simulate.latent_linear}
\alias{latent_simulate.latent_linear}
\title{Simulate outcome}
\usage{
\method{latent_simulate}{latent_linear}(obj, theta, data)
}
\description{
Simulates outcomes from a linear model,
see \link{latent_simulate} for details.
}
