% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_ordinal.r
\name{latent_simulate.latent_ordinal}
\alias{latent_simulate.latent_ordinal}
\title{Simulate outcome}
\usage{
\method{latent_simulate}{latent_ordinal}(obj, theta, data)
}
\description{
Simulates outcomes from an ordinal model,
see \link{latent_simulate} for details.
}
