% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_class.r
\name{latent_simulate}
\alias{latent_simulate}
\title{Simulate outcome from a model}
\usage{
latent_simulate(obj, ...)
}
\arguments{
\item{obj}{A model object.}

\item{theta}{A list of model parameters.}

\item{data}{A data frame with covariates for the simulation.}
}
\value{
An array with simulated outcomes.
}
\description{
Simulates the outcome \code{y} for a model with given
}
