% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log.lik.R
\name{log.lik}
\alias{log.lik}
\title{Log-likelihood function}
\usage{
\method{log}{lik}(data, params, event)
}
\arguments{
\item{data}{matrix of counts, with rows equal to the number of observations and columns equal to the number of categories}

\item{params}{vector of the alpha, beta, and gamma parameters}

\item{event}{vector of labels with one entry per row of data, where each entry identifies the event to which the corresponding row of data belongs}
}
\value{
total log-likelihood of the data under the current parameter values
}
\description{
Log-likelihood of the data under a Poisson distribution with log-linear model for the mean
}

