% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lca.R
\name{print.lca}
\alias{print.lca}
\title{Latent Class Analysis.}
\usage{
lca(data, model = rep("multinomial", ncol(data)), nclasses = 2L,
control = list(opt = "lbfgs", rstarts = 30L, cores = 1L))
}
\arguments{
\item{model}{Character vector with the model for each item.}
}
\value{
Stuff:
\item{df}{Degrees of freedom}
}
\description{
Print the information of latent class models.
}
\details{
None.
}
\references{
None yet.
}
