% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lclass_diag.R
\name{print.lclassd}
\alias{print.lclassd}
\title{Print Method for Latent Class Diagnostics}
\usage{
\method{print}{lclassd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lclassd"} returned by \code{lclass_diag()}.}

\item{...}{Additional arguments passed to \code{print.default} or \code{print.data.frame}.}
}
\description{
Print Method for Latent Class Diagnostics
}
