% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_llca.R
\name{print.llca_plots}
\alias{print.llca_plots}
\title{Print mixed latent class profile plots}
\usage{
\method{print}{llca_plots}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"llca_plots"}.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Redraws the base R plots stored in a mixed latent class profile object.
}
