% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_latent.R
\name{se.latent}
\alias{se.latent}
\title{Variances-Covariance Matrix for Latent Objects}
\usage{
\method{se}{latent}(fit, type = "information", parameters = NULL, digits = 4L, ...)
}
\description{
Variances-Covariance Matrix for Latent Objects
}
