% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_multistage.R
\name{se.multistage}
\alias{se.multistage}
\title{Standard Errors for Latent Objects with multistage estimation.}
\usage{
\method{se}{multistage}(fit, type = "information", parameters = NULL, digits = 4L, ...)
}
\description{
Standard Errors for Latent Objects with multistage estimation.
}
