% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lca_model.R
\name{setup_lca}
\alias{setup_lca}
\title{Get the default model for Latent Class Analysis.}
\usage{
setup_lca(data, item, nclasses, model, control)
}
\arguments{
\item{data}{data.frame or matrix of response.}

\item{nclasses}{Number of latent classes.}

\item{item}{Character vector with the model for each item.}

\item{model}{List of parameter labels. See 'details' for more information.}

\item{constraints}{Should the model be checked for identification? Defaults to TRUE.}
}
\value{
List with the following objects:
\item{none}{.}
\item{none}{.}
}
\description{
Get the default model for Latent Class Analysis.
}
\details{
\code{get_lca} generates the model for the probability of belonging
to the classes and the conditional response probabilities. These models may
be modified by the user to set equality constraints or to fix parameters.
}
\references{
None yet.
}
