% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfactor.R
\name{simfactor}
\alias{simfactor}
\title{Simulate factor structures with misspecification errors.}
\usage{
simfactor(nfactors = 5, nitems = 6, loadings = "medium",
crossloadings = 0, correlations = 0,
estimator = "minres", fit = "rmsr", misfit = 0,
error_method = "cudeck", efa = FALSE,
ngenerals = 0, loadings_g = "medium", correlations_g = 0,
pure = FALSE,
lambda = NULL, Phi = NULL, Psi = NULL)
}
\arguments{
\item{nfactors}{Number of factors.}

\item{nitems}{Number of items per factor.}

\item{loadings}{Loadings' magnitude on the factors: "low", "medium" or "high". Defaults to "medium".}

\item{crossloadings}{Magnitude of the cross-loadings among the group factors. Defaults to 0.}

\item{correlations_g}{Correlation among the general factors. Defaults to 0.}

\item{correlations}{Correlation among the factors. Defaults to 0.}

\item{estimator}{estimator used to generate population error: "minres" or "ml".}

\item{fit}{Fit index to control the population error.}

\item{misfit}{Misfit value to generate population error.}

\item{error_method}{Method used to control population error: c("yuan", "cudeck"). Defaults to "cudeck".}

\item{efa}{Reproduce the error with EFA or CFA. Defaults to FALSE (CFA).}

\item{ngenerals}{Number of general factors.}

\item{loadings_g}{Loadings' magnitude on the general factors: "low", "medium" or "high". Defaults to "medium".}

\item{pure}{Fix a pure item on each general factor. Defaults to FALSE.}

\item{lambda}{Custom loading matrix. If Phi is NULL, then all the factors will be correlated at the value given in correlations.}

\item{Phi}{Custom Phi matrix. If lambda is NULL, then Phi should be conformable to the loading matrix specified with the above arguments.}

\item{Psi}{Custom Psi matrix.}
}
\value{
List with the following objects:
\item{lambda}{Population loading matrix.}
\item{Phi}{Population factor correlation matrix.}
\item{Psi}{Population covariance matrix between the errors.}
\item{R}{Model correlation matrix.}
\item{R_error}{Model correlation matrix with misspecification errors.}
\item{uniquenesses}{Population uniquenesses.}
\item{delta}{Minimum of the loss function that correspond to the misfit value.}
}
\description{
Simulate factor and bifactor structures with crossloadings, correlated factors, and more.
}
\details{
\code{simfactor} generates bi-factor and generalized bifactor patterns with cross-loadings, pure items and
correlations among the general and group factors. When \code{crossloading} is different than 0, one cross-loading
is introduced for an item pertaining to each group factor. When \code{pure} is TRUE, one item loading of each group
factor is removed so that the item loads entirely on the general factor. To maintain the item communalities
constant upon these modifications, the item loading on the other factors may shrunk (if adding cross-loadings)
or increase (if setting pure items).

Loading magnitudes may range between 0.3-0.5 ("low"), 0.4-0.6 ("medium") and 0.5-0.7 ("high"). Custom ranges can be supplied as vectors (i.e., c(0.2, 0.5))
}
\references{

}
