% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{tensor_prod}
\alias{tensor_prod}
\title{Tensor product}
\usage{
tensor_prod(x, y)
}
\arguments{
\item{x, y}{Two vectors of the same length, given either as plain arrays
or as column \code{matrix} objects.}
}
\value{
The tensor product of \code{x} and \code{y} as a square
\code{matrix}.
}
\description{
Forms the tensor product between two vectors.
}
