% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{tensor_square}
\alias{tensor_square}
\title{Tensor square}
\usage{
tensor_square(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The tensor product (see \link{tensor_prod}) between x and itself.
}
\description{
Forms the tensor product of a vector with itself.
}
