% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\name{values}
\alias{values}
\title{Example dataset: values}
\format{
A data frame with 467 rows and 13 variables:
\describe{
  \item{x}{An integer from 1 to 5}
  \item{y}{A lowercase letter}
}
}
\source{
Generated example
}
\usage{
values
}
\description{
A dataset from the poLCA R package with 216 rows and 4 columns.
All the 4 columns are dichotomous items.
}
\keyword{datasets}
