% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfast.R
\name{cfast}
\alias{cfast}
\title{Confirmatory factor analysis.}
\usage{
cfast(
  data,
  target,
  targetphi = NULL,
  targetpsi = NULL,
  cor = "pearson",
  estimator = "uls",
  missing = "pairwise.complete.cases",
  nobs = NULL,
  group = NULL,
  invariance = "metric",
  control = NULL,
  positive = FALSE
)
}
\description{
Confirmatory factor analysis.
}
