% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfast.R
\name{cfast}
\alias{cfast}
\title{Confirmatory factor analysis.}
\usage{
cfast(
  data,
  model = NULL,
  lambda = NULL,
  phi = NULL,
  psi = NULL,
  cor = "pearson",
  estimator = "uls",
  rotate = NULL,
  missing = "pairwise.complete.obs",
  nobs = NULL,
  group = NULL,
  invariance = "none",
  control = NULL,
  std.lv = FALSE,
  positive = FALSE,
  do.fit = TRUE
)
}
\description{
Confirmatory factor analysis.
}
