% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfast.R
\name{cfast}
\alias{cfast}
\title{Fit a Confirmatory Factor Analysis (CFA) model with lavaan syntax.}
\usage{
cfast(data, model = NULL, cor = "pearson", estimator = "ml",
group = NULL, sample.cov = NULL, nobs = NULL,
missing = "pairwise.complete.obs", W = NULL, std.lv = FALSE,
positive = FALSE, do.fit = TRUE, control = NULL)
}
\description{
Fit a Confirmatory Factor Analysis (CFA) model with lavaan syntax.
}
\examples{

\dontrun{
# The famous Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit <- cfast(model = HS.model, data = HolzingerSwineford1939)
summary(fit, fit.measures = TRUE)
}

}
