% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_llca.R
\name{ci.llca}
\alias{ci.llca}
\title{Confidence intervals}
\usage{
ci(fit)
}
\arguments{
\item{fit}{model fitted with lca.}

\item{confidence}{Coverage of the confidence interval.}
}
\value{
List with the following objects:
}
\description{
Compute confidence intervals
}
\details{
Compute confidence intervals.
}
\references{
None yet.
}
