% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_derivs.R
\name{constraints_derivs}
\alias{constraints_derivs}
\title{Derivatives of Parameter Constraints}
\usage{
constraints_derivs(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric matrix containing derivatives of transformation-induced parameter
constraints.
}
\description{
\code{constraints_derivs()} is a generic function for obtaining derivatives
of parameter constraints introduced by model transformations.
}
\details{
Compute derivatives associated with constraints on transformed parameters.
}
\seealso{
\code{\link{jacobian}}, \code{\link{hessian}},
\code{\link{vcov}}
}
