% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.step.R
\name{E.step}
\alias{E.step}
\title{Expectation step}
\usage{
E.step(alpha, beta, gamma, data, event)
}
\arguments{
\item{alpha}{Vector of offsets, length is $d \times p$, where $d$ is the number of unique events and $p$ is the number of FIB species.}

\item{beta}{Vector of slopes, denoting the sensitivity of each FIB species to the level of wastewater contamination. Length is $p$.}

\item{gamma}{Vector of contamination levels, length $n$.}

\item{event}{Vector indicating the event during which each sample was drawn.}

\item{y}{Matrix of observed FIB counts, of dimension $n \times p$. Each row contains the results for one sample, and each column is the counts for one FIB species.}
}
\value{
Matrix of FIB counts, where the censored counts are replaced by their expectations under the current parameters.
}
\description{
Compute the expectations of censored FIB counts, resitricted to be less than the censoring threshold.
}

