% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_llca.R
\name{fitted.llca}
\alias{fitted.llca}
\title{Fitted values for Latent Class Analysis.}
\usage{
fitted(model)
}
\arguments{
\item{model}{Fitted llca object.}
}
\value{
Stuff:
\item{dof}{Degrees of freedom}
}
\description{
Get fitted class membership probabilities for latent class models.
}
\details{
None.
}
\references{
None yet.
}
