% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_llca.R
\name{fitted.llcalist}
\alias{fitted.llcalist}
\title{Fitted Probabilities for Lists of Latent Class Models}
\usage{
\method{fitted}{llcalist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"llcalist"} containing fitted
\code{"llca"} objects.}

\item{...}{Additional arguments passed to \code{fitted.llca()}.}
}
\value{
For an adjusted model with a named \code{structural} component, a numeric
matrix containing the fitted class-membership probabilities from that model.
Otherwise, a list containing one such matrix for each fitted \code{"llca"}
model. The latter result has class \code{"fitted.llcalist"}.
}
\description{
\code{fitted.llcalist()} applies \code{fitted()} to latent class models stored
in an \code{"llcalist"}. For Bakk--Kuha or ML adjustment results containing
elements named \code{measurement} and \code{structural}, only the structural
model is returned because it contains the class-membership regression.
}
\details{
Extract fitted prior class-membership probabilities from an
\code{"llcalist"} object.
}
\references{
None yet.
}
\seealso{
\code{\link{fitted.llca}}, \code{\link{predict.llcalist}}
}
