% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_latent.R
\name{information.latent}
\alias{information.latent}
\title{Information Variance-Covariance Matrix for Latent Models}
\usage{
\method{information}{latent}(fit)
}
\arguments{
\item{fit}{A fitted object inheriting from class \code{"latent"}.}
}
\value{
A list containing the Hessian, variance-covariance matrix, and
  standard errors of the freely estimated parameters.
}
\description{
Information Variance-Covariance Matrix for Latent Models
}
