% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_latent.R
\name{information.latent}
\alias{information.latent}
\title{Hessian and Variance-Covariance Matrix matrix for Latent Models using the
Information method}
\usage{
\method{information}{latent}(fit)
}
\description{
Hessian and Variance-Covariance Matrix matrix for Latent Models using the
Information method
}
