% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_lcfa.R
\name{information.lcfa}
\alias{information.lcfa}
\title{Hessian and Variance-Covariance Matrix matrix for Confirmatory Factor Models
using the Information method}
\usage{
\method{information}{lcfa}(fit)
}
\description{
Hessian and Variance-Covariance Matrix matrix for Confirmatory Factor Models
using the Information method
}
