% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latInspect_lcfa.R
\name{latInspect.lcfa}
\alias{latInspect.lcfa}
\title{Standard Errors}
\usage{
se(fit)
}
\arguments{
\item{fit}{model fitted with lca.}

\item{confidence}{Coverage of the confidence interval.}
}
\value{
List with the following objects:
\item{vcov}{Variance-covariance matrix between the parameters.}
\item{se}{Standard errors.}
\item{SE}{Standard errors in the model list.}
}
\description{
Compute standard errors.
}
\details{
Compute standard errors.
}
\references{
None yet.
}
