% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latInspect_llca.R
\name{latInspect.llca}
\alias{latInspect.llca}
\title{Inspect objects from fitted lca models.}
\usage{
latInspect(fit)
}
\arguments{
\item{fit}{model fitted with lca.}

\item{digits}{Number of digits to print.}
}
\value{
List with one of the following objects:
\item{class}{.}
\item{posterior}{.}
\item{profile}{.}
}
\description{
Inspect objects.
}
\details{
Extract objects from fitted lca object.
}
\references{
None yet.
}
