% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis.}
\usage{
lca(data, model = rep("multinomial", ncol(data)), nclasses = 2L,
control = list(opt = "lbfgs", rstarts = 30L, cores = 1L))
}
\arguments{
\item{data}{data.frame or matrix of response.}

\item{nclasses}{Number of latent classes.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}

\item{itemmodel}{Character vector with the model for each item.}
}
\value{
List with the following objects:
\item{parameters}{The model for the probabilities of the classes.}
\item{f}{Logarithm likelihood at the maximum.}
}
\description{
Estimate latent class models with categorical and continuous data.
}
\details{
\code{lca} estimates models with categorical and continuous data.
}
\references{
None yet.
}
