% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis.}
\usage{
lca(data, item = rep("gaussian", ncol(data)), X = NULL, nclasses = 2L,
    model = NULL, do.fit = TRUE, control = NULL, verbose = TRUE)
}
\arguments{
\item{data}{data frame or matrix.}

\item{nclasses}{Number of latent classes.}

\item{item}{Character vector with the model for each item (i.e., "gaussian" or "multinomial"). Defaults to "gaussian" for all the items.}

\item{X}{Matrix of covariates.}

\item{penalties}{list of penalty terms for the parameters.}

\item{model}{List of parameter labels. See 'details' for more information.}

\item{do.fit}{TRUE to fit the model and FALSE to return only the model setup. Defaults to TRUE.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}
}
\value{
List with the following objects:
\item{parameters}{The model for the logarithm probabilities of the classes.}
\item{f}{Logarithm likelihood at the maximum.}
}
\description{
Estimate latent class models with gaussian and multinomial item models.
}
\details{
\code{lca} estimates models with categorical and continuous data.
}
\references{
None yet.
}
