% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcfa.R
\name{lcfa}
\alias{lcfa}
\title{Fit a Confirmatory Factor Analysis (CFA) model with lavaan syntax.}
\usage{
lcfa(data, model = NULL, estimator = "ml",
ordered = FALSE, group = NULL,
sample.cov = NULL, nobs = NULL,
positive = FALSE, penalties = TRUE,
missing = "pairwise.complete.obs",
std.lv = FALSE, do.fit = TRUE,
message = FALSE, mimic = 'latent',
control = NULL, ...)
}
\arguments{
\item{data}{data frame or matrix.}

\item{model}{lavaan's model syntax.}

\item{estimator}{Available estimators: "ml", "uls", and "dwls". Defaults to "ml".}

\item{ordered}{Logical. Defaults to TRUE.}

\item{group}{String. Name of the variable that splits the data in different groups.}

\item{sample.cov}{Covariance matrix between the items. Defaults to NULL.}

\item{nobs}{Number of observations. Defaults to NULL.}

\item{positive}{Force a positive-definite solution. Defaults to FALSE.}

\item{penalties}{list of penalty terms for the parameters.}

\item{missing}{Method to handle missing data.}

\item{std.lv}{Logical. Provide the parameters of the standardized model. Default is TRUE.}

\item{std.ov}{Logical. Standardize the observed variables before fitting. Default is FALSE.}

\item{acov}{String. "standard" or "robust". Default is "standard".}

\item{meanstructure}{Logical. Estimate the means of the variables. Default is FALSE.}

\item{likelihood}{String. Use N (normal) or N-1 (wishart) in the denominator. Defaults to "normal" for ML and "wishart" otherwise.}

\item{se}{Logical. Compute standard errors. Defaults to TRUE.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}

\item{message}{Logical. Defaults to FALSE.}

\item{do.fit}{TRUE to fit the model and FALSE to return only the model setup. Defaults to TRUE.}

\item{...}{Additional lavaan arguments. See ?lavaan for more information.}
}
\value{
List with the following objects:
\item{version}{Version number of 'latent' when the model was estimated.}
\item{call}{Code used to estimate the model.}
\item{ModelInfo}{Model information.}
\item{Optim}{Output of the optimizer.}
\item{parameters}{Structure with all model parameters.}
\item{transparameters}{Structure with all transformed model parameters.}
\item{loglik}{Logarithm likelihood of the model.}
\item{penalized_loglik}{Logarithm likelihood + logarithm priors of the model.}
}
\description{
Fit a Confirmatory Factor Analysis (CFA) model with lavaan syntax.
}
\details{
\code{lcfa} estimates confirmatory factor models.
}
\examples{

\dontrun{
# The famous Holzinger and Swineford (1939) example
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '

fit <- lcfa(model = HS.model, data = HolzingerSwineford1939)
summary(fit, digits = 3L)
}

}
