% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefa.R
\name{lefa}
\alias{lefa}
\title{Fit an Exploratory Factor Analysis (EFA) model.}
\usage{
lefa(data, nfactors = 1L, model = NULL, estimator = "ml",
ordered = FALSE, group = NULL,
sample.cov = NULL, nobs = NULL,
positive = FALSE, penalties = TRUE,
missing = "pairwise.complete.obs",
std.lv = FALSE, do.fit = TRUE, mimic = 'latent',
control = NULL, ...)
}
\arguments{
\item{data}{data frame or matrix.}

\item{nfactors}{integer. Number of latent variables.}

\item{estimator}{Available estimators: "ml", "uls", and "dwls". Defaults to "ml".}

\item{model}{lavaan's model syntax.}

\item{ordered}{Logical. Defaults to TRUE.}

\item{group}{String. Name of the variable that splits the data in different groups.}

\item{sample.cov}{Covariance matrix between the items. Defaults to NULL.}

\item{nobs}{Number of observations. Defaults to NULL.}

\item{positive}{Force a positive-definite solution. Defaults to FALSE.}

\item{penalties}{list of penalty terms for the parameters.}

\item{missing}{Method to handle missing data.}

\item{std.lv}{Provide the parameters of the standardized model.}

\item{do.fit}{TRUE to fit the model and FALSE to return only the model setup. Defaults to TRUE.}

\item{mimic}{String. Choose the output you want to obtain. Defaults to 'latent'.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}

\item{...}{Additional lavaan arguments. See ?lavaan for more information.}
}
\value{
List with the following objects:
\item{version}{Version number of 'latent' when the model was estimated.}
\item{call}{Code used to estimate the model.}
\item{ModelInfo}{Model information.}
\item{Optim}{Output of the optimizer.}
\item{parameters}{Structure with all model parameters.}
\item{transparameters}{Structure with all transformed model parameters.}
\item{loglik}{Logarithm likelihood of the model.}
\item{penalized_loglik}{Logarithm likelihood + logarithm priors of the model.}
}
\description{
Fit an Exploratory Factor Analysis (EFA) model.
}
\details{
\code{lefa} estimates confirmatory factor models.
}
\examples{

\dontrun{
# The famous Holzinger and Swineford (1939) example

fit <- lefa(data = HolzingerSwineford1939, nfactors = 3L)
summary(fit, digits = 3L)
}

}
