% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpoly.R
\name{lpoly}
\alias{lpoly}
\title{Maximum likelihood estimation of positive definite polychoric correlation matrices.}
\usage{
lpoly(data = NULL,
penalties = TRUE,
do.fit = TRUE,
control = NULL)
}
\arguments{
\item{data}{data frame or matrix with the raw data.}

\item{penalties}{Force a positive-definite solution. Defaults to TRUE.}

\item{do.fit}{TRUE to fit the model and FALSE to return only the model setup. Defaults to TRUE.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}
}
\value{
List with the following objects:
\item{version}{Version number of 'latent' when the model was estimated.}
\item{call}{Code used to estimate the model.}
\item{ModelInfo}{Model information.}
\item{Optim}{Output of the optimizer.}
\item{parameters}{Structure with all model parameters.}
\item{transparameters}{Structure with all transformed model parameters.}
\item{loglik}{Logarithm likelihood of the model.}
\item{penalized_loglik}{Logarithm likelihood + logarithm priors of the model.}
}
\description{
Maximum likelihood estimation of positive definite polychoric correlation matrices.
}
\details{
\code{lpoly} estimates positive-definite polychoric correlation matrices.
}
\examples{

\dontrun{
fit <- lpoly(data = values)
}

}
