% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrotate.R
\name{lrotate}
\alias{lrotate}
\title{Rotate the lambda matrix of an orthogonal factor model.}
\usage{
lrotate(lambda, projection = "oblq", rotation = "oblimin",
 group = NULL, positive = FALSE, penalties = TRUE,
 do.fit = TRUE, control = NULL, ...)
}
\arguments{
\item{lambda}{List, loading matrices for each group.}

\item{projection}{String. Can be "orth", "oblq", or "poblq".}

\item{rotation}{String. Name of the variable that splits the data in different groups.}

\item{group}{String. Name of the variable that splits the data in different groups.}

\item{positive}{Force a positive-definite solution. Defaults to FALSE.}

\item{penalties}{list of penalty terms for the parameters.}

\item{do.fit}{TRUE to fit the model and FALSE to return only the model setup. Defaults to TRUE.}

\item{control}{List of control parameters for the optimization algorithm. See 'details' for more information.}

\item{...}{Additional arguments.}
}
\value{
List with the following objects:
\item{version}{Version number of 'latent' when the model was estimated.}
\item{call}{Code used to estimate the model.}
\item{ModelInfo}{Model information.}
\item{Optim}{Output of the optimizer.}
\item{parameters}{Structure with all model parameters.}
\item{transparameters}{Structure with all transformed model parameters.}
\item{loglik}{Logarithm likelihood of the model.}
\item{penalized_loglik}{Logarithm likelihood + logarithm priors of the model.}
}
\description{
Rotate the lambda matrix of an orthogonal factor model.
}
\details{
\code{lrotate} estimates confirmatory factor models.
}
\examples{

\dontrun{

fit <- lrotate(lambda = , projection = "oblq", rotation = "oblimin")
summary(fit, digits = 3L)
}

}
