% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_llca.R
\name{plot.llca}
\alias{plot.llca}
\title{Plot method for latent class analysis (llca) objects}
\usage{
\method{plot}{llca}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"llca"}, typically from \code{latent}.}

\item{...}{Additional arguments passed to \code{plot_lca_mixed}.}
}
\value{
A list of ggplot objects (as from \code{plot_lca_mixed}).
}
\description{
Plot method for latent class analysis (llca) objects
}
