% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_llca.R
\name{plot.llca}
\alias{plot.llca}
\title{Plot regression coefficients}
\usage{
plot(fit)
}
\arguments{
\item{fit}{model fitted with lca.}

\item{confidence}{Coverage of the confidence interval.}
}
\value{
List with the following objects:
\item{vcov}{Variance-covariance matrix between the parameters.}
\item{se}{Standard errors.}
\item{SE}{Standard errors in the model list.}
}
\description{
Plot the coefficients of latent models.
}
\details{
Obtain a plot with the coefficients of latent models.
}
\references{
None yet.
}
