% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lca_mixed.R
\name{plot_lca_mixed}
\alias{plot_lca_mixed}
\title{Plot mixed LCA results (both Gaussian and multinomial indicators)}
\usage{
plot_lca_mixed(mixed_data, ...)
}
\arguments{
\item{mixed_data}{Output from \code{reshape_lca_mixed}.}

\item{...}{Extra arguments passed to \code{plot_lca_gaussian} and \code{plot_lca_multinomial}.}
}
\value{
A list of ggplot objects, named \code{gaussian} and \code{multinomial}.
}
\description{
Plot mixed LCA results (both Gaussian and multinomial indicators)
}
