% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_llca.R
\name{predict.llca}
\alias{predict.llca}
\title{Predicted values for Latent Class Analysis.}
\usage{
predict(model, new = NULL)
}
\arguments{
\item{model}{Fitted llca object.}

\item{new}{Matrix or data.frame of new predictors.}
}
\value{
Stuff:
\item{dof}{Degrees of freedom}
}
\description{
Get predicted class membership probabilities for latent class models.
}
\details{
None.
}
\references{
None yet.
}
