% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbvr.R
\name{print.lbvr}
\alias{print.lbvr}
\title{Print method for lbvr objects}
\usage{
\method{print}{lbvr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lbvr"}.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\description{
Print method for lbvr objects
}
