% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_latent.R
\name{robust.latent}
\alias{robust.latent}
\title{Hessian and Variance-Covariance Matrix matrix for Latent Models using the
Information method}
\usage{
\method{robust}{latent}(fit)
}
\description{
Hessian and Variance-Covariance Matrix matrix for Latent Models using the
Information method
}
