% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_latent.R
\name{robust.latent}
\alias{robust.latent}
\title{Robust Variance-Covariance Matrix for Latent Models}
\usage{
\method{robust}{latent}(fit)
}
\arguments{
\item{fit}{A fitted object inheriting from class \code{"latent"}.}
}
\value{
The result of \code{information(fit)}.
}
\description{
Default robust method for latent objects without a class-specific robust
estimator. The information covariance matrix is returned.
}
