% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_lcfa.R
\name{robust.lcfa}
\alias{robust.lcfa}
\title{Robust Variance-Covariance Matrix for CFA Models}
\usage{
\method{robust}{lcfa}(fit)
}
\arguments{
\item{fit}{A fitted object inheriting from class \code{"lcfa"}.}
}
\value{
This method currently stops because casewise/patternwise robust
  likelihood scores have not yet been implemented for direct CFA likelihood
  models.
}
\description{
Robust Variance-Covariance Matrix for CFA Models
}
