% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_llca.R
\name{summary.llca}
\alias{summary.llca}
\title{Fit indices}
\usage{
getfit(model)
}
\arguments{
\item{model}{data.frame or matrix of response.}
}
\value{
List with the following fit indices:
\item{AIC}{.}
\item{BIC}{.}
}
\description{
Compute fit indices from any model.
}
\details{
\code{getfit} computes all the fit indices related to a specific model.
}
\references{
None yet.
}
