% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_latent.R
\name{vcov.latent}
\alias{vcov.latent}
\title{Variance-Covariance Matrix for Latent Objects}
\usage{
\method{vcov}{latent}(fit, v, parameters = NULL)
}
\description{
Variance-Covariance Matrix for Latent Objects
}
